Well that is different
Re-architected the i2c "driver" to allow a "handler" to be plugged
in without having to change any of the i2c code. In theory it makes
it possible to put the drive code in a library and then when being
used you just need to write a handler for the type of things you can
do over i2c.

This check-in is 'first light' in the sense that it compiles and echo
is working. Now to get the simulated display code working which should
give me a better feel for how easy or hard it is to write handlers.